# api/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Golang Clean Architecture",
    "description": "Golang Clean Architecture",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/contacts": {
      "post": {
        "tags": [
          "Contact API"
        ],
        "description": "Create new contact",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success create new contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "first_name": {
                          "type": "string"
                        },
                        "last_name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "number"
                        },
                        "updated_at": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Contact API"
        ],
        "description": "Get all contacts",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true
          },
          {
            "name": "page",
            "in": "query",
            "required": false
          },
          {
            "name": "size",
            "in": "query",
            "required": false
          },
          {
            "name": "name",
            "in": "query",
            "required": false
          },
          {
            "name": "phone",
            "in": "query",
            "required": false
          },
          {
            "name": "email",
            "in": "query",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success get list of contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "first_name": {
                            "type": "string"
                          },
                          "last_name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          },
                          "phone": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "number"
                          },
                          "updated_at": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "paging": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "size": {
                          "type": "number"
                        },
                        "total_item": {
                          "type": "number"
                        },
                        "total_page":{
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}